+++ src/Shared/Recruit.Vacancies.Client/Application/CommandHandlers/CreateVacancyCommandHandler.cs
@@ -6,24 +6,24 @@ using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Handlers
+namespace Esfa.Recruit.Vacancies.Client.Application.CommandHandlers
 {
-    public class SubmitVacancyCommandHandler : IRequestHandler<SubmitVacancyCommand>
+    public class CreateVacancyCommandHandler: IRequestHandler<CreateVacancyCommand>
     {
         private readonly IVacancyRepository _repository;
         private readonly IMessaging _messaging;
 
-        public SubmitVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
+        public CreateVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
         {
             _repository = repository;
             _messaging = messaging;
         }
 
-        public async Task Handle(SubmitVacancyCommand message, CancellationToken cancellationToken)
+        public async Task Handle(CreateVacancyCommand message, CancellationToken cancellationToken)
         {
-            await _repository.UpdateAsync(message.Vacancy);
+            await _repository.CreateAsync(message.Vacancy);
 
-            await _messaging.PublishEvent(new VacancySubmittedEvent
+            await _messaging.PublishEvent(new VacancyCreatedEvent
             {
                 SourceCommandId = message.CommandId.ToString(),
                 EmployerAccountId = message.Vacancy.EmployerAccountId,
