+++ src/Shared/Recruit.Vacancies.Client/Application/CommandHandlers/DeleteVacancyCommandHandler.cs
@@ -6,24 +6,24 @@ using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Handlers
+namespace Esfa.Recruit.Vacancies.Client.Application.CommandHandlers
 {
-    public class CreateVacancyCommandHandler: IRequestHandler<CreateVacancyCommand>
+    public class DeleteVacancyCommandHandler : IRequestHandler<DeleteVacancyCommand>
     {
         private readonly IVacancyRepository _repository;
         private readonly IMessaging _messaging;
 
-        public CreateVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
+        public DeleteVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
         {
             _repository = repository;
             _messaging = messaging;
         }
 
-        public async Task Handle(CreateVacancyCommand message, CancellationToken cancellationToken)
+        public async Task Handle(DeleteVacancyCommand message, CancellationToken cancellationToken)
         {
-            await _repository.CreateAsync(message.Vacancy);
+            await _repository.UpdateAsync(message.Vacancy);
 
-            await _messaging.PublishEvent(new VacancyCreatedEvent
+            await _messaging.PublishEvent(new VacancyDeletedEvent
             {
                 SourceCommandId = message.CommandId.ToString(),
                 EmployerAccountId = message.Vacancy.EmployerAccountId,
