+++ src/Shared/Recruit.Vacancies.Client/Application/CommandHandlers/SubmitVacancyCommandHandler.cs
@@ -6,24 +6,24 @@ using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Handlers
+namespace Esfa.Recruit.Vacancies.Client.Application.CommandHandlers
 {
-    public class DeleteVacancyCommandHandler : IRequestHandler<DeleteVacancyCommand>
+    public class SubmitVacancyCommandHandler : IRequestHandler<SubmitVacancyCommand>
     {
         private readonly IVacancyRepository _repository;
         private readonly IMessaging _messaging;
 
-        public DeleteVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
+        public SubmitVacancyCommandHandler(IVacancyRepository repository, IMessaging messaging)
         {
             _repository = repository;
             _messaging = messaging;
         }
 
-        public async Task Handle(DeleteVacancyCommand message, CancellationToken cancellationToken)
+        public async Task Handle(SubmitVacancyCommand message, CancellationToken cancellationToken)
         {
             await _repository.UpdateAsync(message.Vacancy);
 
-            await _messaging.PublishEvent(new VacancyDeletedEvent
+            await _messaging.PublishEvent(new VacancySubmittedEvent
             {
                 SourceCommandId = message.CommandId.ToString(),
                 EmployerAccountId = message.Vacancy.EmployerAccountId,
