+++ src/Shared/Recruit.Vacancies.Client/Application/CommandHandlers/UpdateApprenticeshipProgrammesCommandHandler.cs
@@ -4,7 +4,7 @@ using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Handlers
+namespace Esfa.Recruit.Vacancies.Client.Application.CommandHandlers
 {
     public class UpdateApprenticeshipProgrammesCommandHandler: IRequestHandler<UpdateApprenticeshipProgrammesCommand>
     {
