+++ src/Shared/Recruit.Vacancies.Client/Application/CommandHandlers/UpdateUserCommandHandler.cs
@@ -5,7 +5,7 @@ using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Esfa.Recruit.Vacancies.Client.Application.Handlers
+namespace Esfa.Recruit.Vacancies.Client.Application.CommandHandlers
 {
     public class UpdateUserCommandHandler : IRequestHandler<UpdateUserCommand>
     {
